==> R/cisFeatureModel/6_logit.R <==
# logistic regression

load('train.RData')
library(e1071)
library(class)
library(ROCR)
library(caret)

logitCV <- function(data, formula, n, n.site=1000) {
  # the return value - a list of prediction tables
  pred.list <- list()
  for (i in 1:n) {
    
    all.index <- 1 : nrow(data)
    train.index <- ((i - 1) * n.site + 1) : (i * n.site)
    test.index <- all.index[-train.index]  # all sites except train sites
    train <- data[train.index,]
    test <- data[test.index,]
    #logistic regression
    print(paste("Running CV:", i))
    logit<-glm(formula, train, family=binomial)
    score<-predict(logit, test, type="response")
    label<-test$class
    beta<-test$class
    # pred is the predicted class
    pred <- score
    pred[pred >= 0.5] = 1; pred[pred < 0.5] = 0
    pred <- as.factor(pred)
    
    pred.table <- data.frame(label = label, beta = test$beta, pred = pred, score = score)
    pred.list[[i]] <- pred.table
  }
  
  pred.list
}
var.remove <- nearZeroVar(features)
logit.data <- features[,-var.remove]
logit.formula <- paste(intersect(colnames(logit.data), colnames(features)[c(7, 9, seq(14, 141))]), collapse = '+')
logit.formula <- paste0('as.factor(class)~', logit.formula)
pred.logit <- logitCV(logit.data, logit.formula, 10, 1750)


evalRF <- function(pred.list, n.fold) {
    acc.v <- NULL
    auc.v <- NULL
    spec.v <- NULL
    sens.v <- NULL
    mcc.v <- NULL
    tp.v <- NULL
    fp.v <- NULL
    tn.v <- NULL
    fn.v <- NULL
    for (i in seq(n.fold)) {
        ROC <- prediction(pred.list[[i]]$score, pred.list[[i]]$label)

        auc.pref <- performance(ROC, "auc")
        auc <- auc.pref@y.values[[1]]
        auc.v <- c(auc.v, auc)

        pred <- pred.list[[i]]$pred
        label <- pred.list[[i]]$label
        tp <- as.numeric(length(which(pred==1 & label==1)))
        tn <- as.numeric(length(which(pred==0 & label==0)))
        fp <- as.numeric(length(which(pred==1 & label==0)))
        fn <- as.numeric(length(which(pred==0 & label==1)))
        tp.v <- c(tp.v, tp)
        tn.v <- c(tn.v, tn)
        fp.v <- c(fp.v, fp)
        fn.v <- c(fn.v, fn)

        spec <- tn / (tn + fp)
        spec.v <- c(spec.v, spec)

        sens <- tp / (tp + fn)
        sens.v <- c(sens.v, sens)

        acc <- (tp + tn) / (tp + fp + tn + fn)
        acc.v <- c(acc.v, acc)

        mcc <- (tp * tn - fp * fn) / sqrt((tp + fn) * (tp + fp) * (tn + fp) * (tn + fn))
        mcc.v <- c(mcc.v, mcc)
    }
    data.frame(acc.v, auc.v, spec.v, sens.v, mcc.v, tp.v, fp.v, tn.v, fn.v)
}


eval.logit <- evalRF(pred.logit, 10)
save(pred.logit, eval.logit, file='logitRes.RData')
write.table(eval.logit, file='logitRes.csv', sep=',', row.names=T, col.names=F, quote=F)
